transcriber_service/requirements.txt
speechrecognition==3.10.0
kafka-python==2.0.2
elasticsearch==8.15.1